lxd: Move `ExpandInstanceConfig` and `ExpandInstanceDevices` from `lxd/db` to `lxd/instance/instancetype` package
