+++ lxd/db/instances_test.go
@@ -265,8 +265,8 @@ func TestInstanceList(t *testing.T) {
 
 	var instances []db.InstanceArgs
 	err = c.InstanceList(context.TODO(), func(dbInst db.InstanceArgs, p api.Project) error {
-		dbInst.Config = db.ExpandInstanceConfig(dbInst.Config, dbInst.Profiles)
-		dbInst.Devices = db.ExpandInstanceDevices(dbInst.Devices, dbInst.Profiles)
+		dbInst.Config = instancetype.ExpandInstanceConfig(dbInst.Config, dbInst.Profiles)
+		dbInst.Devices = instancetype.ExpandInstanceDevices(dbInst.Devices, dbInst.Profiles)
 
 		instances = append(instances, dbInst)
